Improve code quality: split monoliths, deduplicate shaders, extract phases
- Split main.rs handle_worker_results (781 lines) into 9 per-outcome handlers
- Extract shared GLSL into shader_common.rs, eliminating ~70% duplication
  between gpu_renderer.rs and globe_radar_renderer.rs
- Split canvas.rs (2,064 lines) into canvas_overlays, canvas_interaction,
  and canvas_inspector modules
- Extract worker_api.rs ingest phases into reusable functions in
  ingest_phases.rs, reducing worker_api from 1,766 to 1,340 lines
- Genericize decode_worker.rs message handlers with extract_pending_context
  helper, eliminating 5x duplicated boilerplate
